requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-telegram-bot==21.3
openai==1.30.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.130.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
Pillow==10.3.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
aiosqlite==0.20.0
tenacity==8.3.0
structlog==24.1.0
python-dateutil==2.9.0.post0
trafilatura==1.10.0
